MetricWave_API/requirements.txt
flask~=3.0.3
flask-sqlalchemy
mysql-connector-python
python-dotenv~=1.0.1
passlib~=1.7.4
flasgger~=0.9.7.1
itsdangerous~=2.2.0
